+++ includes/service-providers/constant_contact/class-newspack-newsletters-constant-contact-sdk.php
@@ -785,6 +785,20 @@ final class Newspack_Newsletters_Constant_Contact_SDK {
 			if ( isset( $data['taggings'] ) ) { // Using isset and not empty because this can be an empty array.
 				$body['taggings'] = $data['taggings'];
 			}
+			if ( isset( $body['update_source'], $data['email'] ) ) {
+				$body['email_address']['address'] = $data['email'];
+				$existing_contact                 = $this->get_contact( $data['email'] );
+				if ( $existing_contact && ! \is_wp_error( $existing_contact ) ) {
+					try {
+						// If we are updating to an existing email address, delete the old one and update the existing one.
+						$this->request( 'DELETE', 'contacts/' . $contact->contact_id );
+					} catch ( Exception $e ) {
+						Newspack_Newsletters_Logger::log( 'Error deleting contact during upsert with email ' . $data['email'] . ': ' . $e->getMessage() );
+					}
+				}
+				$contact       = $existing_contact;
+				$email_address = $data['email'];
+			}
 		}
 
 		try {
@@ -794,9 +808,8 @@ final class Newspack_Newsletters_Constant_Contact_SDK {
 				[ 'body' => wp_json_encode( $body ) ]
 			);
 		} catch ( Exception $e ) {
-			return new WP_Error( 'newspack_newsletter_error_upserting_contact', $e->getMessage() );
+			return new WP_Error( 'newspack_newsletters_constant_contact_api_error', $e->getMessage() );
 		}
-
 		return $this->get_contact( $email_address );
 	}