+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp-cached-data.php
@@ -7,7 +7,7 @@
 
 defined( 'ABSPATH' ) || exit;
 
-use DrewM\MailChimp\MailChimp;
+use Newspack_Newsletters_Mailchimp_Api as Mailchimp;
 
 /**
  * Mailchimp cached class data
@@ -375,7 +375,7 @@ final class Newspack_Newsletters_Mailchimp_Cached_Data {
 				<?php
 				echo esc_html(
 					__(
-						'Newspack Newsletters is having trouble to fetch data from Mailchimp Audiences. Please visit Newsletters > Settings for more details.',
+						'Newspack Newsletters is having trouble fetching data from Mailchimp Audiences. Please visit Newsletters > Settings for more details.',
 						'newspack_newsletters'
 					)
 				);