+++ src/newsletter-editor/store.js
@@ -17,6 +17,7 @@ import { addQueryArgs } from '@wordpress/url';
 /**
  * Internal dependencies
  */
+import { isManualESP, isSupportedESP } from './utils';
 import { getServiceProvider } from '../service-providers';
 
 /**
@@ -117,6 +118,10 @@ export const updateNewsletterDataError = error =>
 
 // Dispatcher to fetch newsletter data from the server.
 export const fetchNewsletterData = async postId => {
+	if ( ! isSupportedESP() || isManualESP() ) {
+		return;
+	}
+
 	const isRetrieving = coreSelect( STORE_NAMESPACE ).getIsRetrieving();
 	if ( isRetrieving ) {
 		return;
@@ -148,6 +153,10 @@ export const fetchNewsletterData = async postId => {
 
 // Dispatcher to fetch any errors from the most recent sync attempt.
 export const fetchSyncErrors = async postId => {
+	if ( ! isSupportedESP() || isManualESP() ) {
+		return;
+	}
+
 	const isRetrieving = coreSelect( STORE_NAMESPACE ).getIsRetrieving();
 	if ( isRetrieving ) {
 		return;
@@ -170,6 +179,10 @@ export const fetchSyncErrors = async postId => {
 
 // Dispatcher to fetch send lists and sublists from the connected ESP and update the newsletterData in store.
 export const fetchSendLists = debounce( async ( opts, replace = false ) => {
+	if ( ! isSupportedESP() || isManualESP() ) {
+		return [];
+	}
+
 	updateNewsletterDataError( null );
 	try {
 		const { name } = getServiceProvider();